re-factor dsa_make_key() etc.
