+++ tests/dsa_test.c
@@ -59,13 +59,6 @@ static char *hex_q = "AA5BD7F4E5062413E58835CA00C7A635716194C5";
 static char *hex_x = "9936E5E4E9FB28BE91F5065FE8C935B3F5D81FC5";
 static char *hex_y = "5316B0FBBF598A5E5595C14FAC43B80853E6CF0D9223FAB184595239BFCBF22D383ADD935205497E2B12C46173E36F54BD96E5A7AAA95A58A4B767D2C0BDC81EB13A124F98C005EF395D6ABAB70B3BD8B795DD796EA2D28473470388B464D9B9B84FF1C934BBF97366F57C2E11FEC331E60838596781EB6D4127D70D74AFA035";
 
-/* private key - raw decimal numbers */
-static char *dec_g = "41834149751984197912953436480983170533071735026506895442815002322147255782590882063707309354781506433716654796985480894012184326029507913813728323760888731712844346877576824916725534905000120412305763983626878322597033839508975868744887842375259196379140567488975525420966465471602331600963525846901216912348";
-static char *dec_p = "138366127874251453574215823372867983172559870428080754538874699342292548213873551009389476481395012375639515165022292709776266658812209612126692196557051247870332681145778007636026326219557730049370214260237710845864302921876857532769906463917243319959886290876544710558897185626634470575981605420411381006287";
-static char *dec_q = "972576611327916959546542817054443329226761409733";
-static char *dec_x = "874699854785640347852049895863914110365034094533";
-static char *dec_y = "58346825863862115220306694056113472976936045407556113559931032566376300411053620606958863235131122432665794570437845128216268156672161823000705623178942581094085367656740608001229642983928728905397237964247962716781137229394844332774819193277135681825866994604976120931444766148118918668354923664000689348661";
-
 /* The public part of test_dsa.key in SubjectPublicKeyInfo format */
 static const unsigned char openssl_pub_dsa[] = {
   0x30, 0x82, 0x01, 0xb6, 0x30, 0x82, 0x01, 0x2b, 0x06, 0x07, 0x2a, 0x86,
@@ -107,11 +100,13 @@ static const unsigned char openssl_pub_dsa[] = {
   0xeb, 0x6d, 0x41, 0x27, 0xd7, 0x0d, 0x74, 0xaf, 0xa0, 0x35
 };
 
-static int dsa_compat_test(void)
+static int _dsa_compat_test(void)
 {
-  dsa_key key;
+  dsa_key key = LTC_DSA_KEY_INITIALIZER;
   unsigned char tmp[1024], buf[1024];
   unsigned long x, len;
+  unsigned char key_parts[5][256];
+  unsigned long key_lens[5];
 
   DO(dsa_import(openssl_priv_dsa, sizeof(openssl_priv_dsa), &key));
 
@@ -135,7 +130,22 @@ static int dsa_compat_test(void)
   dsa_free(&key);
 
   /* try import private key from raw hexadecimal numbers */
-  DO(dsa_import_radix(16, hex_p, hex_q, hex_g, hex_x, hex_y, &key));
+  for (x = 0; x < 5; ++x) {
+     key_lens[x] = sizeof(key_parts[x]);
+  }
+  DO(radix_to_bin(hex_p, 16, key_parts[0], &key_lens[0]));
+  DO(radix_to_bin(hex_q, 16, key_parts[1], &key_lens[1]));
+  DO(radix_to_bin(hex_g, 16, key_parts[2], &key_lens[2]));
+  DO(radix_to_bin(hex_y, 16, key_parts[3], &key_lens[3]));
+  DO(radix_to_bin(hex_x, 16, key_parts[4], &key_lens[4]));
+
+  DO(dsa_set_pqg(key_parts[0], key_lens[0],
+                 key_parts[1], key_lens[1],
+                 key_parts[2], key_lens[2],
+                 &key));
+  DO(dsa_set_key(key_parts[3], key_lens[3],
+                 key_parts[4], key_lens[4],
+                 &key));
   len = sizeof(buf);
   DO(dsa_export(buf, &len, PK_PRIVATE | PK_STD, &key));
   if (len != sizeof(openssl_priv_dsa) || memcmp(buf, openssl_priv_dsa, len)) {
@@ -144,32 +154,18 @@ static int dsa_compat_test(void)
   }
   dsa_free(&key);
 
-  /* try import private key from raw decimal numbers */
-  DO(dsa_import_radix(10, dec_p, dec_q, dec_g, dec_x, dec_y, &key));
-  len = sizeof(buf);
-  DO(dsa_export(buf, &len, PK_PRIVATE | PK_STD, &key));
-  if (len != sizeof(openssl_priv_dsa) || memcmp(buf, openssl_priv_dsa, len)) {
-     fprintf(stderr, "DSA private export failed to match dsa_import_radix(10, ..)\n");
-     return 1;
-  }
-  dsa_free(&key);
-
   /* try import public key from raw hexadecimal numbers */
-  DO(dsa_import_radix(16, hex_p, hex_q, hex_g, NULL, hex_y, &key));
+  DO(dsa_set_pqg(key_parts[0], key_lens[0],
+                 key_parts[1], key_lens[1],
+                 key_parts[2], key_lens[2],
+                 &key));
+  DO(dsa_set_key(key_parts[3], key_lens[3],
+                 NULL, 0,
+                 &key));
   len = sizeof(buf);
   DO(dsa_export(buf, &len, PK_PUBLIC | PK_STD, &key));
   if (len != sizeof(openssl_pub_dsa) || memcmp(buf, openssl_pub_dsa, len)) {
      fprintf(stderr, "DSA public export failed to match dsa_import_radix(16, ..)\n");
-     return 1;
-  }
-  dsa_free(&key);
-
-  /* try import public key from raw decimal numbers */
-  DO(dsa_import_radix(10, dec_p, dec_q, dec_g, NULL, dec_y, &key));
-  len = sizeof(buf);
-  DO(dsa_export(buf, &len, PK_PUBLIC | PK_STD, &key));
-  if (len != sizeof(openssl_pub_dsa) || memcmp(buf, openssl_pub_dsa, len)) {
-     fprintf(stderr, "DSA public export failed to match dsa_import_radix(10, ..)\n");
      return 1;
   }
   dsa_free(&key);
@@ -182,12 +178,14 @@ int dsa_test(void)
    unsigned char msg[16], out[1024], out2[1024], ch;
    unsigned long x, y;
    int stat1, stat2;
-   dsa_key key, key2;
+   dsa_key key = LTC_DSA_KEY_INITIALIZER;
+   dsa_key key2 = LTC_DSA_KEY_INITIALIZER;
 
-   dsa_compat_test();
+   _dsa_compat_test();
 
    /* make a random key */
-   DO(dsa_make_key(&yarrow_prng, find_prng("yarrow"), 20, 128, &key));
+   DO(dsa_generate_pqg(&yarrow_prng, find_prng("yarrow"), 20, 128, &key));
+   DO(dsa_make_key_ex(&yarrow_prng, find_prng("yarrow"), &key));
 
    /* verify it */
    DO(dsa_verify_key(&key, &stat1));
